Refactor test coverage reporting commands in circle ci
Why:
Instead of generating several reports, this will generate a single
lcov report for all packages and projects in the repository.

The downside is that we will have to rethink how envs for other
client projects will be handled. At this point in time i can only
hope that we can always mock out the env.ts file such that we
would not need the actual envs when running tests,o
 ever.