Fix QA bugs
Hide the inherit setting action for settings that are inherited. You can know whether a setting is inherited if the locationId is not the same as the selected locationId
Some settings are showing the locationId on the inherited from column even if the location is valid and has a valid location label.
Some settings do not show the inherited from a location yet they are inherited